fix nvd provider to not wipe out existing results on incremental update
