+++ tests/unit/providers/nvd/test_manager.py
@@ -2,8 +2,7 @@ import json
 
 import pytest
 
-from vunnel import result, workspace
-from vunnel.providers import nvd
+from vunnel import workspace
 from vunnel.providers.nvd import manager
 
 
@@ -31,27 +30,3 @@ def test_parser(tmpdir, helpers, mock_data_path, mocker):
     actual_vulns = [v for v in subject.get(None)]
 
     assert expected_vulns == actual_vulns
-
-
-@pytest.mark.parametrize(
-    "mock_data_path,expected_written_entries",
-    [
-        ("test-fixtures/single-entry.json", 1),
-    ],
-)
-def test_provider_schema(helpers, mock_data_path, expected_written_entries, mocker):
-    workspace = helpers.provider_workspace_helper(name=nvd.Provider.name())
-    mock_data_path = helpers.local_dir(mock_data_path)
-
-    with open(mock_data_path) as f:
-        json_dict = json.load(f)
-
-    c = nvd.Config()
-    c.runtime.result_store = result.StoreStrategy.FLAT_FILE
-    p = nvd.Provider(root=workspace.root, config=c)
-    p.manager.api.cve = mocker.Mock(return_value=[json_dict])
-
-    p.update(None)
-
-    assert expected_written_entries == workspace.num_result_entries()
-    assert workspace.result_schemas_valid(require_entries=expected_written_entries > 0)
